+++ docs/js/raqdps.js
@@ -22,6 +22,15 @@ let startTime = null
 let endTime = null
 let defaultTIme = null
 let currentTime = null;
+let dateIsLocal = true;
+const dateOptions = {
+  year: 'numeric',
+  month: '2-digit',
+  day: '2-digit',
+  hour: '2-digit',
+  minute: '2-digit',
+  timeZoneName: 'short'
+};
 
 let layers = [
     new ol.layer.Tile({
@@ -89,9 +98,14 @@ function updateLayers() {
   layers[1].getSource().updateParams({'TIME': currentTime.toISOString().split('.')[0]+"Z"});
 }
 
+// Dispalying current map time
 function updateInfo() {
   let el = document.getElementById('info');
-  el.innerHTML = `Time / Heure: ${currentTime.toISOString().substr(0, 16)+"Z"}`
+  if (dateIsLocal) {
+    el.innerHTML = `Time/Heure: ${currentTime.toLocaleString(navigator.local, dateOptions)}`
+  } else {
+    el.innerHTML = `Time/Heure: ${currentTime.toISOString().substr(0, 16)+"Z"}`
+  }
 }
 
 // Disable/enable buttons depending on the state of the map
@@ -257,6 +271,11 @@ function exportMapFunction(e) {
   map.renderSync();
 };
 
+function switchDateFormat(e) {
+  dateIsLocal = !dateIsLocal
+  updateInfo()
+}
+
 let fastBackwardButton = document.getElementById('fast-backward');
 fastBackwardButton.addEventListener('click', fastBackward, false);
 
@@ -274,6 +293,9 @@ fastForwardButton.addEventListener('click', fastForward, false);
 
 let exportButton = document.getElementById('exportmap');
 exportButton.addEventListener('click', exportMapFunction, false);
+
+let dateInfo = document.getElementById('info');
+dateInfo.addEventListener('click', switchDateFormat, false);
 
 // Initialize the map
 function initMap() {