+++ CHANGELOG.md
@@ -20,6 +20,7 @@ MSC's public documentation uses [Semantic Versioning](https://semver.org/) for i
     * Add scale line on interactive maps
     * Add a label to distinguish between observation/extrapolation on weather radar interactive maps
     * Add current conditions interactive map to City Page Weather documentation
+    * Add a switch option on datetime to change between local time/UTC on interactve maps
 * bug fixes:
     * None